Move some duplicated/shared card logic between failure/event into 1 place
