+++ frontend/src/components/event/index.js
@@ -1,24 +1,17 @@
 import React, { Component, PropTypes } from 'react'
 import { connect } from 'react-redux'
-import moment from 'moment'
 
 import { showEventOverview } from 'actions/event-overview'
-import style from 'components/event/style'
+import cardStyle, {
+  formatTime,
+  formattedEventType
+} from 'components/event/card-style'
 
 import {Card, CardHeader, CardTitle} from 'material-ui/Card'
 import EventsIcon from 'material-ui/svg-icons/communication/call-received'
 import EventOverviewDialog from 'components/event-overview-dialog'
 import EventRetryDialog from 'components/event-retry-dialog'
 import { green200 } from 'material-ui/styles/colors'
-
-const EVENT_TIME_FORMAT = 'h:mm:ss a'
-const EMPTY_EVENT_TYPE = '<no type>'
-
-export function formatEventTime (eventTime) {
-  if (!eventTime) return null
-  const eventTimeDate = new Date(eventTime)
-  return moment(eventTimeDate).format(EVENT_TIME_FORMAT)
-}
 
 export class Event extends Component {
   static get propTypes () {
@@ -48,18 +41,17 @@ export class Event extends Component {
     return (
       <Card
         className='event'
-        style={style.card}
+        style={cardStyle.card}
         onClick={() => this.showOverview()}>
         <CardHeader
           className='event-header'
           avatar={<EventsIcon className='event-icon' color={green200} />}
-          titleStyle={style.cardHeader.title}
-          subtitleStyle={style.cardHeader.subtitle}
-          title={formatEventTime(this.props.event.event_time)}
+          titleStyle={cardStyle.cardHeader.title}
+          title={formatTime(this.props.event.event_time)}
           subtitle={this.props.event.topic}/>
         <CardTitle
-          titleStyle={style.cardTitle}
-          title={this.formatedEventType()}/>
+          titleStyle={cardStyle.cardTitle}
+          title={formattedEventType(this.props.event.event_type)}/>
         <EventOverviewDialog event={this.props.event} />
         <EventRetryDialog event={this.props.event} />
       </Card>
@@ -68,10 +60,6 @@ export class Event extends Component {
 
   showOverview () {
     this.props.onShowOverview(this.props.event)
-  }
-
-  formatedEventType () {
-    return this.props.event.event_type || EMPTY_EVENT_TYPE
   }
 }
