+++ frontend/src/components/event/index.spec.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import jasmineEnzyme from 'jasmine-enzyme'
 import { mount } from 'enzyme'
-import { Event, formatEventTime } from 'components/event'
+import { Event } from 'components/event'
+import { formatTime } from 'components/event/card-style'
 import getMuiTheme from 'material-ui/styles/getMuiTheme'
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import { Provider } from 'react-redux'
@@ -51,7 +52,7 @@ describe('<Event />', () => {
   })
 
   it('displays event_time formatted', () => {
-    expect(component.text()).toMatch(formatEventTime(props.event.event_time))
+    expect(component.text()).toMatch(formatTime(props.event.event_time))
   })
 
   it('displays topic', () => {
