--- frontend/src/components/event/style.js
@@ -1,16 +1,0 @@
-export default {
-  card: {
-    width: '490px',
-    overflow: 'hidden',
-    backgroundColor: '#fff'
-  },
-  cardHeader: {
-    title: {color: '#000', fontWeight: 'lighter'},
-    subtitle: {color: '#000'}
-  },
-  cardTitle: {
-    color: '#000',
-    fontSize: '38px',
-    fontWeight: 'lighter'
-  }
-}
