+++ frontend/src/components/failure/index.spec.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import jasmineEnzyme from 'jasmine-enzyme'
 import { mount } from 'enzyme'
-import { Failure, formatTime } from 'components/failure'
+import { Failure } from 'components/failure'
+import { formatTime } from 'components/event/card-style'
 import getMuiTheme from 'material-ui/styles/getMuiTheme'
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import { Provider } from 'react-redux'
